requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==20.3.0
autopep8==2.0.2
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
cs50==9.2.5
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
Flask>=1.0,<2.3
Flask-Session==0.4.1
frozenlist==1.3.3
helpers==0.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.10
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.7
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==4.0.0
pycodestyle==2.10.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
SQLAlchemy==1.4.46
sqlparse==0.4.4
termcolor==2.3.0
tqdm==4.65.0
urllib3==1.26.7
virtualenv==20.24.7
Werkzeug==2.3.3
yarl==1.9.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0